Docs: Add ITRocket Decentralisation Analitic (#1610)
We are glad to announce our new service - Celestia decentralization analytics. 

As part of the analytics we provide:
- a decentralization map
- decentralization indexes by regions, countries and hosting providers
- voting power distribution index

Analytics is provided for different types of nodes: all nodes, RPC nodes, and validator nodes. 

We hope that this tool will help improve the decentralization and reliability of the network.

Please note that the tool is still under development, and analytics and indicators will be further improved. You are welcome to share any ideas and feedback.